fix: deduplicate shared enums into runtime
