+++ packages/runtime-tags/src/common/types.ts
@@ -52,6 +52,49 @@ export enum AccessorChar {
   PreviousAttributes = "~",
 }
 
+export enum NodeType {
+  Element = 1,
+  Text = 3,
+  Comment = 8,
+  DocumentFragment = 11,
+}
+
+// Reserved Character Codes
+// 0-31 [control characters]
+// 34 " [double quote]
+// 39 ' [single quote]
+// 92 \ [backslash]
+// 96 ` [backtick]
+export enum WalkCode {
+  Get = 32,
+  Before = 33,
+  After = 35,
+  Inside = 36,
+  Replace = 37,
+  EndChild = 38,
+
+  BeginChild = 47,
+
+  Next = 67,
+  NextEnd = 91,
+
+  Over = 97,
+  OverEnd = 106,
+
+  Out = 107,
+  OutEnd = 116,
+
+  Multiplier = 117,
+  MultiplierEnd = 126,
+}
+
+export enum WalkRangeSize {
+  Next = 20, // 67 through 91
+  Over = 10, // 97 through 106
+  Out = 10, // 107 through 116
+  Multiplier = 10, // 117 through 126
+}
+
 export type Accessor = string | number;
 export type Input = Record<string, unknown>;
 export type Context = Record<string, unknown>;
