+++ packages/runtime-tags/src/dom/walker.ts
@@ -1,5 +1,4 @@
-import type { Scope } from "../common/types";
-import { NodeType } from "./dom";
+import { type Scope, NodeType, WalkCode, WalkRangeSize } from "../common/types";
 import { createScope } from "./scope";
 
 export const walker = /* @__PURE__ */ document.createTreeWalker(document);
@@ -17,42 +16,6 @@ export const walker = /* @__PURE__ */ document.createTreeWalker(document);
 //    - Before must be done before walking into the node
 //    - Next would walk back in the node we just walked Out of
 //  - A component must assume the walker is on its first node, and include instructions for walking to its assumed nextSibling
-
-// Reserved Character Codes
-// 0-31 [control characters]
-// 34 " [double quote]
-// 39 ' [single quote]
-// 92 \ [backslash]
-// 96 ` [backtick]
-export enum WalkCode {
-  Get = 32,
-  Before = 33,
-  After = 35,
-  Inside = 36,
-  Replace = 37,
-  EndChild = 38,
-
-  BeginChild = 47,
-
-  Next = 67,
-  NextEnd = 91,
-
-  Over = 97,
-  OverEnd = 106,
-
-  Out = 107,
-  OutEnd = 116,
-
-  Multiplier = 117,
-  MultiplierEnd = 126,
-}
-
-export enum WalkRangeSize {
-  Next = 20, // 67 through 91
-  Over = 10, // 97 through 106
-  Out = 10, // 107 through 116
-  Multiplier = 10, // 117 through 126
-}
 
 export function trimWalkString(walkString: string): string {
   let end = walkString.length;
