+++ packages/translator-tags/src/core/for.ts
@@ -5,6 +5,7 @@ import {
   getTagDef,
 } from "@marko/babel-utils";
 import { types as t } from "@marko/compiler";
+import { WalkCode } from "@marko/runtime-tags/common/types";
 import { isOutputHTML } from "../util/marko-config";
 import analyzeAttributeTags from "../util/nested-attribute-tags";
 import { mergeReferences } from "../util/references";
@@ -80,7 +81,7 @@ export default {
         extra: { isOnlyChild },
       } = tag.node;
       if (!isOnlyChild) {
-        walks.visit(tag, walks.WalkCode.Replace);
+        walks.visit(tag, WalkCode.Replace);
         walks.enterShallow(tag);
       }
       if (isOutputHTML()) {
